jupiter-controller-update: Include .service, handle failures
Because the binary is a bash script, we need to make sure a SIGTERM
sent to it causes the firmware tool to similarly get killed, which is
surprisingly not the default case when you SIGTERM a bash script.

Also, timeout after 90s, because we're seeing this hang and cannot
reproduce.

make a half-hearted attempt to reset the controller boards when this
happens -- if it is failing at one of the two enumerate steps, this
should save you, otherwise you'll need to reboot and try again.