format-sdcard: add a quick format option (default)
We now skip block discards by default during format. A full discard
helps cards that have previously been filled, but that isn't the case
for a new card.

If a card has been previously filled and it is getting reformatted this
could be useful, but it would also be useful for someone that filled
their card and then deleted some data without formatting. Hence the
proper place to address discard seems to be at runtime.

We leave the door open to re-enable a "full" format (i.e. with discards)
by specifying the --full flag.