Workaround for an occasional failure to format sdcards
There is currently a race between mkfs and the lsblk command in that is
part of the sdcard-mount service. This results in lsblk sometimes
returning null as the fstype for a partition.

We workaround this issue by polling at the end of the format script
until lsblk actually notices the change. I was hoping that partprobed
would be the thing we needed to synchronize things correctly without the
need for a retry loop, but that doesn't seem to fix it in all cases. And
maybe partprobed only makes the bug less common because it introduces a
delay itself.

We also add better logging for the sdcard-mount failures and return a
special error code when we run into mount errors. That way the steamui
can show a more specific error message for these cases.