Add format-device.sh script and PolicyKit helper
This allows formatting storage devices other than /dev/mmcblk0.

It is identical to the old format-sdcard.sh script but has a new
'--device /dev/XXX' parameter and sanity checks to ensure that we're
actually trying to format a removable device and not the system drive.

In order to keep backward compatibility the old script is now a
wrapper that calls the new one with --device /dev/mmcblk0