jupiter-biosupdate: Don't auto-apply BIOS updates in OOBE builds
This primarily lets USB repair/recovery images not attempt to
auto-update things and reboot out of themselves, they have an explicit
bios update step as part of performing a re-image, and doing so is
unexpected.  But in general, an OOBE build with a newer bios should
just hold off and do its day1 update.