automount: Fix url variable in send_steam_url()
The local url variable is a leftover from commit ed65881 but it is
never used where it is defined. The 'else' branch of the conditional
does print a url, which happens to work because it uses the value set
by the callers (but not passed to the function).

This patch gets rid of that inconsistency and simplifies the code by
moving all instances of urlencode from the callers to send_steam_url()