format-device: add support for setting filesystem owner
This adds support for a --owner argument, that accepts a uid:gid, to be
passed directly to mkfs.ext4's root_owner extended option.

The reason for this parameter is to allow setting filesystem ownership
on creation. Without this, the created filesystem is owned by the user
who invoked this script (often root), and would be RO or unusable for
any user who may have triggered root to run this script.

I thought an argument to set the ownership was better than hard-coding
some uid:gid (e.g. "1000:1000"), since that assumption might not always
work out. This does default to 1000:1000 if --owner isn't specified, so
that it's backwards compatible with callers that don't set --owner.

Fixes https://gitlab.steamos.cloud/jupiter/tasks/-/issues/832

I looked at several options for fixing 832, but ultimately decided that
this was a good approach. It's tempting to fix this in udisks2 somehow,
but a filesystem created/owned by root is still owned by root when it's
mounted by a user (e.g. without the 'as-user'), and I'm not sure we
should modify permissions in an already existing filesystem on some disk
a user plugs in