format-device: Remove check for hotplug devices
Our check uses lsblk to query the 'hotplug' flag. This ultimately
comes from the 'removable' flag from the kernel, which is known to be
unreliable:

   https://github.com/KDE/solid/commit/6d260195cf75604d835235d2a1b02166ee8b514a

Since we have additional checks in place we can live without this one,
so let's just remove it.

Fixes https://github.com/ValveSoftware/SteamOS/issues/1064