workaround bluetooth auto-reconnect failures by restarting service on resume
This attempts to work around issues I've observed where bluetooth
headsets do not reconnect (or take tens of seconds / a few minutes) to
reconnect when the Deck resumes from suspend. Restarting the bt service
is a bit heavy-handed, but it makes the reconnect on resume MUCH more
reliable in my testing thus far, so it seems like a reasonable
workaround for now until the real root cause can be determined.

Potentially also helps with user reported issues like:

https://steamcommunity.com/app/1675200/discussions/2/3824162683825778743/
https://www.reddit.com/r/SteamDeck/comments/u3krot/bluetooth_headphones_not_reconnecting_after/