automount: Make idmapped mount if the fs is not owned by the deck user
The deck user needs write access to an external drive in order to be
able to use it properly. This is not a problem if the filesystem is
created by us because we set the ownership of the root directory at
format time.

Filesystems formatted elsewhere can be owned by a different user. In
this case we can make an idmapped mount so the actual owner of the
root directory is mapped to the deck user. With this the filesystem
can be used without having to use chown.

Although this should cover the most common scenarios we'll still have
problems if the filesystem is owned by different users (for example if
the root and steamapps directories have different ownership).

https://gitlab.steamos.cloud/jupiter/tasks/-/issues/916