automount: run fsck before mounting an external drive
If an ext4 filesystem was not unmounted cleanly then the journal is
automatically replayed on the next mount and this is usually enough to
guarantee its consistency.

However if the filesystem is known to have errors then the kernel
simply prints a warning and carries on.

This patch tries to repair the filesystem before mounting it. If
the errors can be corrected without user intervention this succeeds
silently, otherwise we refuse to mount it and report an error to the
user instead.

If the filesystem is clean then this only replays the journal (if
necessary) and adds almost no overhead.

https://gitlab.steamos.cloud/jupiter/tasks/-/issues/928