automount: Run fsck with -y and try to mount the fs ro if it has errors
If a filesystem has errors that cannot be corrected without
user intervention we are simply reporting the problem to Steam.
Unfortunately the user does not currently receive any kind of visual
feedback so it appears as if the external drive was not detected.

With this patch we run fsck with -y so it tries to correct all errors
and not just the simple ones. If this still fails we try to mount
filesystem in read-only mode, so it can still be used if it's not too
broken.

https://gitlab.steamos.cloud/jupiter/tasks/-/issues/661
(cherry picked from commit 389dd5f6ee452cbf719785559ec3ddeee661fe39)