automount: Add a separate script to handle udev add/remove events
The steamos-automount service can be started from two different
places:

- The udev add / remove events from the block subsystem.
- The format-device script.

We are using a lock file in order to prevent attempts to mount the
device while it's being formatted.

The problem is that it is the automount script itself that tries to
take the lock, leaving the service in a failed state when the lock
cannot be taken. This happens for example when an external drive is
formatted: the format script holds the lock and all the udev events
that are emitted in the process result in errors that end up in the
journal. In addition to that, the format script needs to release the
lock before finishing so the drive can be finally mounted.

This is not a very clean solution so this patch moves the handling of
the udev events to a separate script and only tries to mount the drive
when no one else is using it. This also allows us to filter the udev
events based on additional criteria, as we'll do in the next patch.

This patch also moves the lock handling to the common-functions
file and adds a sanity check to ensure that the device name is not
malformed.