automount: Use TimeMediaDetected to decide whether to mount a drive
When an external drive is inserted or removed it triggers udev events
that we use to mount or unmount it.

Those udev events can however be triggered by other actions:

- When we alter the partition table.
- When we try to eject the drive with /usr/bin/eject.
  (this is harder to reproduce with util-linux 2.39; earlier versions
   were causing interactions with Plasma's eject functionality)

These make our automount script think that the SD card was physically
removed and inserted, so it ends up being remounted again.

In order to handle this case we can use the TimeMediaDetected property
exposed by UDisks. This tells us when the drive was originally
detected and it is not affected by this eject operation.

With this, if the partition (udev "add" event) appears long after the
drive is detected then it means that the event did not happen because
the media was just plugged in but because of some other reason.
In this case no automount should happen.

This is the same approach that was used in GNOME:

https://gitlab.gnome.org/GNOME/gvfs/-/commit/b4800b987b4a8423a52306c9aef35b3777464cc5

Corresponding KDE bug:

https://bugs.kde.org/show_bug.cgi?id=467751

One special case is the 'cold plug' scenario, when the media was
already inserted at boot time. GNOME needs special code to handle this
case (see https://bugzilla.redhat.com/show_bug.cgi?id=813069), but
since our script runs early at boot time and is run directly by udev
when the drive is detected we don't need to do anything else.

Fixes https://gitlab.steamos.cloud/jupiter/tasks/-/issues/806