automount: Ensure that the deck user can write to an external drive
The deck user needs write access to an external drive in order to be
able to use it properly. This is not a problem if the filesystem is
created by us because we set the ownership of the root directory at
format time.

Filesystems formatted elsewhere can have different permissions and
ownership. In the most basic case Steam only needs to be able to
write to the root directory so it can create libraryfolder.vdf and
steamapps. Because of this if we change the permissions of the root
directory we would cover the most common scenarios without making
invasive changes (we'll still have problems if the filesystem already
contains a steamapps folder owned by a different user).

Making the root directory of the filesystem world-writable is also
the approach taken by e.g. the KDE partition manager when formatting
external drives.

https://gitlab.steamos.cloud/jupiter/tasks/-/issues/916