jupiter-controller-update-wrapper: add plymouth wrapper
Add a small wrapper that gets executed instead of the original in the
systemd service. It serves to check if update is required and if so,
enable plymouth and draw smooth progress bar using it.

Once complete, switch back to the normal boot mode (static image atm) in
order to pass that along to next piece in the boot chain.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>