jupiter-controller-update: confirm firmware file(s)
Currently we don't validate that the firmware file(s) are present. As
result, if a firmware update is needed although the file is not there
(like on my EV2 unit), we'll trigger the update process only to error
out.

This get particularly messy with the upcoming wrapper, which will
ensure plymouth is running as needed. Without this change we'll start
plymouth, start drawing the progress and exit midway.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>