Reinstate the EV2 firmware
While the unit is no longer supported, we still check and report if an
update is needed. In addition, my unit is an EV2 one is running a newer
version and miss-behaving badly.

Reinstate the file, it's negligible in size and unlikely to change in
the future. Unblocking me and others who have such an old unit.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>