Work around misdetection of some SanDisk microSD cards
Some models of SanDisk microSD cards are occassionally misdetected by
the kernel. When this happens they appear with the SD032 name and with
a capacity of 30.6 MB.

The user can solve this problem by removing and reinserting the card,
but it can also be solved at the software level by rebinding the
sdhci-pci device, forcing a rescan of the microSD card.

The only problem with this solution is that a bug in Linux < 6.1.72
can crash the kernel. This was fixed upstream in commit 1036f69e2513.

https://github.com/ValveSoftware/SteamOS/issues/1339