grub: /etc/default/grub-steamos was out of date
History:
 - Intended to be the repository for steamos specific grub settings
 - Applied after /etc/default/grub
 - grub-mkconfig was unaware so was bypassed in the image builder yaml

Imported changes:
 - Added:
   - log_buf_len
   - amdgpu.lockup_timeout
   - amdgpu.sched_hw_submission
   - fbcon=rotate
 - Altered
   - fbcon=vc:2-6 → fbcon=vc:4-6

We're about to make grub-mkconfig respect the settings in this file,
but before we do that we need to make it reflect the current kernel
command line settings we actually use.

This have been imported from the holo/holo monorepo.

They were/are in the disk.yaml file there (and should be
removed from there before this is merged into the monorepo)

        #