format-device.sh: run 'udevadm settle' before mkfs.ext4
This script runs parted to create a new partition table and then runs
mkfs.ext4 to format its only partition.

However it is possible that mkfs.ext4 starts before the partition is
visible in /dev, and the existing 'sync' call does not always prevent
that. So this patch replaces it with 'udevadm settle'.

There is also no need to call sync after the partition is formatted
since mkfs.ext4 already does that.