automount: ignore OS partitions
The automounter is intended for removable¹ media devices.

We don't want it to pick up OS partitions as this can block the OS
update from running (by mounting the other slot's /var, for example)

This affects systems installed on /dev/sdX based systems (eg running
off a USB device, or a non-NVME SSD, or even spinning rust).

Since we have the OS partitions UUIDs in our partset data we can
ignore those partitions in the automount registration script.

1: removable here means "which can be removed or detached", which is
subtly and annoyingly not the same as "has the removable attribute".